\name{write.gmoney}
\alias{write.gmoney}
\title{write.gmoney}
\usage{
write.gmoney(dat, file)
}
\description{
Writes a file with "g$" as the separator. Will save as a ".g$" file if no extension specified. Includes fun console output as well!
}
\examples{
write.gmoney(dat, file="mydata")
Creates File named mydata.g$
}
