\name{write.nathan}
\alias{write.nathan}
\title{write.nathan}
\usage{
write.nathan(dat, file)
}
\description{
Writes a file with "nathan" as the separator. Will save as a ".nathan" file if no extension specified. Includes fun console output as well!
}
\examples{
write.nathan(dat, file="mydata")
Creates File named mydata.nathan
}
